pyproject.toml
[project]
name = "namingpaper"
version = "0.2.7"
description = "CLI tool to rename, organize, and search academic papers using AI-extracted metadata"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "DanTsai0903"}]
keywords = ["pdf", "academic", "papers", "rename", "metadata", "ai", "cli", "library", "search", "organize"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Utilities",
]

dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pymupdf>=1.26.7",
    "httpx>=0.27.0",
]

[project.urls]
Homepage = "https://github.com/DanTsai0903/namingpaper"
Repository = "https://github.com/DanTsai0903/namingpaper"
Issues = "https://github.com/DanTsai0903/namingpaper/issues"

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
gemini = ["google-generativeai>=0.5.0", "Pillow>=10.0.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0"]
claude = [
    "anthropic>=0.30.0",
]

[project.scripts]
namingpaper = "namingpaper.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/namingpaper"]

[tool.hatch.build.targets.editable]
packages = ["src/namingpaper"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    "ignore::DeprecationWarning:importlib._bootstrap",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]
